pyproject.toml
[tool.poetry]
name = "dropshipping"
version = "0.1.0"
description = ""
authors = ["ShevaPython <shevadotka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
django-allauth = {extras = ["socialaccount"], version = "^0.63.3"}


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
